Implement requested changes, put threading back
https://github.com/RagtagOpen/bidwire/pull/48#pullrequestreview-46136803
