+++ bidwire/notifiers/boston_bid_notifier.py
@@ -1,7 +1,6 @@
 from .notifier_utils import make_bid_item_body, make_doc_item_body
 from .base_notifier import BaseNotifier
 from bid import Bid
-from notice import Notice
 
 
 class CityOfBostonNotifier(BaseNotifier):
@@ -15,14 +14,3 @@ class CityOfBostonNotifier(BaseNotifier):
 
     def make_item_body(self, item):
         return make_bid_item_body(item)
-
-
-class CityOfBostonNoticeNotifier(BaseNotifier):
-    def get_site(self):
-        return Notice.Site.BOSTON
-
-    def make_item_body(self, item):
-        return make_doc_item_body(item)
-
-    def get_listings_pre_text(self, items_length):
-        return "We have found {} new notices since we last sent you an update:".format(items_length)
