+++ bidwire/scrapers/boston_bid_scraper.py
@@ -1,17 +1,12 @@
 import logging
 import re
-from datetime import datetime
-from datetime import date as dtdate
-from queue import Queue
-from concurrent.futures import ProcessPoolExecutor, ThreadPoolExecutor
+from concurrent.futures import ThreadPoolExecutor
 import scrapelib
 
 from lxml import html
 
 from .base_scraper import BaseScraper
 from bid import Bid, get_new_identifiers
-from document import Document
-from utils import execute_parallel
 
 # Logger object for this module
 log = logging.getLogger(__name__)
@@ -20,12 +15,11 @@ compiled_reg_exp = re.compile("bids\.asp\?ID=(\d+)")
 
 
 class CityOfBostonScraper(BaseScraper):
-    results_url = "https://www.cityofboston.gov/purchasing/bid.asp"
-    details_url = "https://www.cityofboston.gov/purchasing/bids.asp"
+    RESULTS_URL = "https://www.cityofboston.gov/purchasing/bid.asp"
+    DETAILS_URL = "https://www.cityofboston.gov/purchasing/bids.asp"
 
     def __init__(self, threads=1, processes=4):
         self.scraper = scrapelib.Scraper()
-        self.proc_executor = ProcessPoolExecutor(processes)
         self.thread_executor = ThreadPoolExecutor(threads)
 
     def scrape_bids(self, session):
@@ -43,7 +37,7 @@ class CityOfBostonScraper(BaseScraper):
         Arguments:
         session -- database session to use for persisting items
         """
-        bid_page = self.scraper.get(self.results_url)
+        bid_page = self.scraper.get(self.RESULTS_URL)
         bid_ids = self.scrape_results_page(bid_page.content)
         log.info("Found bid ids: {}".format(bid_ids))
         new_ids = get_new_identifiers(session, bid_ids, self.get_site())
@@ -51,7 +45,7 @@ class CityOfBostonScraper(BaseScraper):
         session.bulk_save_objects(bids)
 
     def get_bid_page(self, bid_id):
-            return bid_id, self.scraper.get(self.details_url, params={'ID': bid_id}).content
+            return bid_id, self.scraper.get(self.DETAILS_URL, params={'ID': bid_id}).content
 
     def scrape(self, session):
         self.scrape_bids(session)
@@ -106,61 +100,3 @@ class CityOfBostonScraper(BaseScraper):
             items=items,
             site=self.get_site().name
         )
-
-
-class NoticesScraper(BaseScraper):
-    notices_url = "https://www.boston.gov/public-notices"
-
-    def __init__(self, threads=1, processes=4):
-        self.scraper = scrapelib.Scraper()
-        self.proc_executor = ProcessPoolExecutor(processes)
-        self.thread_executor = ThreadPoolExecutor(threads)
-
-    def scrape_desc(self, href):
-        desc_page = self.scraper.get('https://www.boston.gov' + href)
-        tree = html.fromstring(desc_page.content)
-        try:
-            return html.tostring(tree.xpath('//div["field-item even"=@class]')[0]).decode()
-        except:
-            # the description is a list
-            return html.tostring(tree.xpath('//div["body"=@class]/ol')[0]).decode()
-
-
-    def scrape_notice_div(self, div):
-        title_a = div.xpath(".//div['n-li-t'=@class]/a")[0]
-        year, month, day_and_start, end = div.xpath("//span['dc:date'=@property]/@content")[0].split('-')
-        day, start = day_and_start.split('T')
-        date = dtdate(int(year), int(month), int(day))
-        posted_candidates = div.xpath("//span['dl-d'=@class]")
-        for cand in posted_candidates:
-            if cand.text is None:
-                continue
-            if re.match('\d\d/\d\d/\d\d\d\d - \d:\d\d[ap]m', cand.text):
-                posted = datetime.strptime(cand.text, '%m/%d/%Y - %I:%M%p')
-                break
-        else:
-            raise ValueError("Couldn't get time of post")
-        return Document(
-            url='https;//www.boston.gov' + title_a.attrib['href'],
-            title=title_a.attrib['title'],
-            site=Document.Site.BOSTON.name,
-            description=self.scrape_desc(title_a.attrib['href'])
-        )
-
-    def get_site(self):
-        return Bid.Site.CITYOFBOSTON
-
-    def scrape_notices_page(self, content):
-        tree = html.fromstring(content)
-        notice_divs = tree.xpath('//div["g g--m0 n-li"=@class]')
-        log.info("Found {} notices".format(len(notice_divs)))
-        for div in notice_divs:
-            yield self.scrape_notice_div(div)
-
-    def scrape_notices(self):
-        notices_page = self.scraper.get(self.notices_url)
-        return self.scrape_notices_page(notices_page.content)
-
-    def scrape(self, session):
-        session.bulk_save_objects(self.scrape_notices())
-        session.commit()
